backend/agentcore_runtimes/content/requirements.txt
bedrock-agentcore==1.4.6
strands-agents==1.30.0
boto3>=1.34
pydantic>=2.0,<3.0
cachetools==7.0.5
